+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-test/test/localStorage.spec.ts
@@ -1,3 +1,4 @@
+import fs from "fs";
 import puppeteer from "puppeteer";
 import { Screenshot, createFolder, setupCredentials, enterCredentials, storagePoller, ONE_SECOND_IN_MS } from "../../../../../e2eTestUtils/TestUtils";
 import { BrowserCacheUtils } from "../../../../../e2eTestUtils/BrowserCacheTestUtils";
@@ -5,9 +6,7 @@ import { LabApiQueryParams } from "../../../../../e2eTestUtils/LabApiQueryParams
 import { AzureEnvironments, AppTypes } from "../../../../../e2eTestUtils/Constants";
 import { LabClient } from "../../../../../e2eTestUtils/LabClient";
 import { msalConfig as aadMsalConfig, request as aadTokenRequest } from "../authConfigs/localStorageAuthConfig.json";
-import { clickLoginPopup, clickLoginRedirect, waitForReturnToApp } from "./testUtils";
-import fs from "fs";
-import {getBrowser, getHomeUrl} from "../../testUtils";
+import { clickLoginPopup, clickLoginRedirect, waitForReturnToApp, getBrowser, getHomeUrl } from "./testUtils";
 
 const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots/localStorageTests`;
